Project made to work on local machine
It was much simpler to do the project on the local machine in the time frame given to us. This does not contain the implementation of the duplicate image checker. It creates new folders every minute and saves images of people not wearing masks in it every 10 seconds or so.